switched back to geoapi 2.1-SNAPSHOT; reverted GeographicBoundingBox
git-svn-id: https://svn.osgeo.org/geotools/geotools/trunk@25001 e5c1c795-43da-0310-a71f-fac65c449510

+++ gt/modules/library/referencing/src/main/java/org/geotools/geometry/GeneralEnvelope.java
@@ -193,10 +193,10 @@ public class GeneralEnvelope implements Envelope, Cloneable, Serializable {
     public GeneralEnvelope(final GeographicBoundingBox box) {
         ensureNonNull("box", box);
         ordinates = new double[] {
-            box.getWestBoundLongitude().doubleValue(),
-            box.getSouthBoundLatitude().doubleValue(),
-            box.getEastBoundLongitude().doubleValue(),
-            box.getNorthBoundLatitude().doubleValue()
+            box.getWestBoundLongitude(),
+            box.getSouthBoundLatitude(),
+            box.getEastBoundLongitude(),
+            box.getNorthBoundLatitude()
         };
         crs = DefaultGeographicCRS.WGS84;
     }
